+++ framework/PageCells/LabelsCell.tsx
@@ -7,17 +7,22 @@ type LabelsWithLinksProps = {
   labels?: never;
   labelsWithLinks: LabelWithLink[];
   numLabels?: number;
+  wrapLabels?: boolean;
 };
 
 type LabelsProps = {
   labels: string[];
   labelsWithLinks?: never;
   numLabels?: number;
+  wrapLabels?: boolean;
 };
 
 export function LabelsCell(props: LabelsProps | LabelsWithLinksProps) {
   return (
-    <LabelGroup numLabels={props.numLabels ?? 999} style={{ flexWrap: 'nowrap' }}>
+    <LabelGroup
+      numLabels={props.numLabels ?? 999}
+      style={props.wrapLabels ? undefined : { flexWrap: 'nowrap' }}
+    >
       {props.labels
         ? props.labels.map((label) => <Label key={label}>{label}</Label>)
         : props.labelsWithLinks.map((labelWithLink) => (
